Make all metrics double instead of int

+++ statsd.net.shared/Messages/Counter.cs
@@ -8,10 +8,10 @@ namespace statsd.net.shared.Messages
 {
   public sealed class Counter : StatsdMessage
   {
-    public int Value { get; set; }
+    public double Value { get; set; }
     public float? SampleRate { get; set; }
 
-    public Counter(string name, int value)
+    public Counter(string name, double value)
     {
       if (value < 0)
       {
@@ -23,7 +23,7 @@ namespace statsd.net.shared.Messages
       MessageType = MessageType.Counter;
     }
 
-    public Counter(string name, int value, float sampleRate)
+    public Counter(string name, double value, float sampleRate)
     {
       if (value < 0)
       {
