+++ statsd.net/Framework/TimedGaugeAggregatorBlockFactory.cs
@@ -23,7 +23,7 @@ namespace statsd.net.Framework
       IIntervalService intervalService,
       ILog log)
     {
-      var gauges = new ConcurrentDictionary<string, int>();
+      var gauges = new ConcurrentDictionary<string, double>();
       var root = rootNamespace;
       var ns = String.IsNullOrEmpty(rootNamespace) ? "" : rootNamespace + ".";
 
@@ -45,7 +45,7 @@ namespace statsd.net.Framework
           if (removeZeroGauges)
           {
             // Get all zero-value gauges
-            int placeholder;
+            double placeholder;
             var zeroGauges = 0;
             for (int index = 0; index < items.Length; index++)
             {
